Simplify docstring stripping for the shellcraft command line tool.
We can get away with the code being simpler as long as we ensure that
all of the docstrings come after the string "Example".  Most of them
did, and I fixed the ones that did not.

Some of the Example blocks came before the Arguments block, I moved
those up so they wouldn't be stripped.

Additionally, some of the docstrings are actually indented. For example,
compare `i386.stackhunter` to `i386.mov`.  The `textwrap` module has a
utility function to uniformly dedent text.  I've added this to make
everything left-aligned.