fix: Use `os.makedirs()` to create cache dir
This change ensures that the cache directory will be created along with
any missing intermediate path components. This fixes a misbehaviour
where the cache would not be created if, e.g. `$HOME/.cache` does not
already exist. `exist_ok=True` is used to ensure that we allow the use
of an existing cache directory while still capturing other `OSError`s
such as permissions errors when we attempt to create it.

If we fail to create the cache directory (likely due to permission
issues) or it ends up being unwritable we return `None` and expect other
modules to handle this situation gracefully.