Prevent IndexError on ui.options
Hello

I sugest to modify this ===> elif k in min(tuple(string.digits):...
into this =============> elif k in min(tuple(string.digits), tuple(map(str, range(len(opts)))), key=len):...

The problem is, when the list of options is less of 9 elements, if i press a digit who is not in len(opts), the program breaks, cause IndexError. One possible solution is taking the shortest tuple between tuple(string.digits) and a new tuple(map(str, range(len(opts))))).

It is not a big probleme, but today, when i tried to do something with this function, i was able to identify this problem, and i fixed it with this quick trick. 
I not good in english, and is my first pull request. I appreciate all critiques =)

Tanks for your work