Force `ELF.path` to be of type string (#2174)
This commit forces the `path` class variable of the `ELF` class to be of
type string, even if a byte string argument is passed. This also makes
the implementation consistent with the documentation, which states that
`ELF.path` is of type `str`.

Fixes #2166

Cherry-picked from 60af3a26d0b2b4777e7050727c0140451cf747c5

Signed-off-by: James Raphael Tiovalen <jamestiotio@gmail.com>